Added functionality for file delete

+++ server/routes/handlers/files.js
@@ -177,19 +177,36 @@ let files = {
   // TODO: Get file by Id, then delete file from folder based on sourceUrl
   // Be carefull not to delete parent folder(s)
   // fs.unlink()
-  delete: (request, reply) => {
-    models.File.destroy({
-        'where': {
-          'id': request.params.id
-        }
-      })
-      .then((file) => {
-        if (file) {
-          reply().code(200);
-        } else {
-          reply().code(404);
-        }
-      });
+	delete: (request, reply) => {
+    models.File.find({
+      'where': {
+        'id': request.params.id
+      }
+    }).then((file) => {
+      if (!file.locationUrl || file.locationUrl.slice(-1) === '/' || file.locationUrl.indexOf('.') < 0) {
+        reply(Boom.notAcceptable('File object is missing a proper locationUrl property'));
+      } else {
+				let locationUrl = __dirname + '/../../..' + env.uploadPath + file.locationUrl;
+        models.File.destroy({
+            'where': {
+              'id': request.params.id
+            }
+          })
+          .then((file) => {
+            if (file) {
+							fse.unlink(locationUrl, (err) => {
+								if (err) {
+									reply(Boom.badRequest('Error deleting file.'));
+									return;
+								}
+								reply().code(200);
+							});
+            } else {
+              reply().code(404);
+            }
+          });
+      }
+    });
   }
 };
